Break out code a bit more
