+++ packages/multi-test/src/lib.rs
@@ -1,5 +1,8 @@
 mod balance;
 mod bank;
 mod handlers;
+mod wasm;
 
 pub use crate::bank::{Bank, SimpleBank};
+pub use crate::handlers::Router;
+pub use crate::wasm::{Contract, ContractWrapper, WasmRouter};
